# be/docs/openapi/cmms-api.yaml
openapi: 3.1.0
info:
  title: CMMS API
  version: 1.0.0
  description: >
    Computerized Maintenance Management System REST API.
    Modules: Auth, Dashboard, Equipment, Work Orders, Schedule, Technicians,
    Spare Parts, Reports, Alerts, Settings. Roles: Admin, Manager, Technician.
servers:
  - url: http://localhost:5000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/equipment:
    get:
      tags: [Equipment]
      summary: List equipment (paginated, filterable)
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/EquipmentStatus' } }
        - { name: categoryId, in: query, schema: { type: string, format: uuid } }
        - { name: locationId, in: query, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/SortBy' }
        - { $ref: '#/components/parameters/SortDir' }
      responses:
        '200':
          description: Paginated equipment
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/equipment.json#/listPaginated
    post:
      tags: [Equipment]
      summary: Create equipment (Admin, Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEquipment' }
            examples:
              default:
                externalValue: ./examples/equipment.json#/create
      responses:
        '201':
          description: Created
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/equipment.json#/detail
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/equipment/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [Equipment]
      summary: Get equipment by id
      responses:
        '200':
          description: Equipment
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/equipment.json#/detail
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Equipment]
      summary: Update equipment (Admin, Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEquipment' }
            examples:
              default:
                externalValue: ./examples/equipment.json#/update
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Equipment]
      summary: Soft-delete equipment (Admin, Manager)
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/reports/kpi:
    get:
      tags: [Reports]
      summary: MTTR, MTBF, OEE impact, overdue rate
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: KPI report
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/reports.json#/kpi
  /api/reports/cost:
    get:
      tags: [Reports]
      summary: Cost breakdown (labor vs parts)
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Cost report
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/reports.json#/cost
  /api/reports/equipment-history/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [Reports]
      summary: Maintenance timeline for an equipment
      responses:
        '200':
          description: Equipment history
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/reports.json#/equipmentHistory
  /api/reports/downtime:
    get:
      tags: [Reports]
      summary: Downtime by equipment
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: Downtime report
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/reports.json#/downtime
  /api/alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      parameters:
        - { name: acknowledged, in: query, schema: { type: boolean } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Paginated alerts
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/alerts.json#/list
  /api/alerts/{id}/acknowledge:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    patch:
      tags: [Alerts]
      summary: Acknowledge an alert
      responses:
        '200':
          description: Acknowledged
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/alerts.json#/acknowledgeResponse
  /api/alerts/{id}/create-wo:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    post:
      tags: [Alerts]
      summary: Create a work order from an alert (Admin, Manager)
      responses:
        '201':
          description: Work order created
          content:
            application/json:
              examples:
                ok:
                  externalValue: ./examples/alerts.json#/createWoFromAlertResponse
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    SortBy:
      name: sortBy
      in: query
      schema: { type: string }
    SortDir:
      name: sortDir
      in: query
      schema: { type: string, enum: [asc, desc] }
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            default:
              externalValue: ./examples/errors.json#/forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            default:
              externalValue: ./examples/errors.json#/notFound
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            default:
              externalValue: ./examples/errors.json#/validation
  schemas:
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
        details:
          type: array
          items: { type: string }
      required: [error, code]
...
    EquipmentStatus:
      type: string
      enum: [Active, UnderMaintenance, Broken, Decommissioned]
...
    CreateEquipment:
      type: object
      required: [code, name, categoryId, locationId, status]
      properties:
        code: { type: string }
        name: { type: string }
        categoryId: { type: string, format: uuid }
        locationId: { type: string, format: uuid }
        manufacturer: { type: string, nullable: true }
        installDate: { type: string, format: date, nullable: true }
        status: { $ref: '#/components/schemas/EquipmentStatus' }
    UpdateEquipment:
      allOf:
        - $ref: '#/components/schemas/CreateEquipment'
...